[REFACTOR] Certik Audit Recommendation - Exhibit 12
Description:
The declarations contained in lines 54 and 69 do not possess any type of visibility specifier.

Recommendations:
A visibility specifier should be added to denote who has access to the values of these variables
in an on-chain sense. If the variables are declared as “internal” or “private”, the underscore
naming convention should be kept. Otherwise, the underscore should be removed from the
declaration names.