[REFACTOR] Certik Audit Recommendation - Exhibit 21
Description:
The code block between 200-202 accesses the “length” member of the “_dividendsSnapshots”
array twice whereas it is also transferred to “memory” on line 206.

Recommendations:
The code block should be moved past the “totalSnapshots” declaration of line 206 and utilize
the in-memory variable rather than the in-storage “length” member. Note that an optimization
step may already be taken here by the compiler and as such, the gas optimization may not be
visible in the transactions conducted.