[REFACTOR] Certik Audit Recommendation - Exhibit 20
Description:
The function “_getOrCreateLatestCycleSnapshot” returns a “DividendsSnapshot” struct stored in
“memory” whereas in both functions that it is called (L190 & L640) it ends up being replacing
itself after being edited (L192 & L656).

Recommendations:
This is very inefficient as we require the copy of the complete “struct” from “storage” to
“memory” and vice versa, even when we affect only a single member of the “struct”. By returning
a “storage” pointer, the assignments on lines 192 and 656 would be unnecessary as the
“storage” itself would be directly affected by the statements in the functions. Additionally, other
areas of the codebase can also be adjusted or removed such as L225-227 and L233-L235. On a
minor note, a type exists in comment line 189.