[CHANGED] Certik Audit - Exhibit 3 (round 3)
[INFORMATIONAL] Description:
The variables in the aforementioned lines concern two “immutable” contract addresses, an “IERC20” address and an “IERC1155” address.

Recommendations:
These variables are never utilized as-is, i.e. in their “address” format. As such, instead of casting them to the corresponding contracts on each invocation it is possible to store them as contract-type addresses on construction.