[CHANGED] Certik Audit - Exhibit 5 (round 3)
[INFORMATIONAL] Description:
The owner is able to withdraw the deposited reward tokens at will via the “withdrawRewardsPool” function.

Recommendations:
This means that, even though the “start” function guarantees enough funds are transferred to the contract to satisfy the reward payouts, the owner would still be able to withdraw them post-start maliciously in case he wants to back-track from providing dividends. This goes against the current paradigm of “funding” the contract on start because the owner can in essence have as many funds as he wishes within the contract.