[CHANGED] Certik Audit - Exhibit 10 (round 3)
[INFORMATIONAL] Description:
Should “currentCycle” be equal to the snapshot’s “startCycle” and the history provided having at least one snapshot within, the “_updateHistory” function will follow a code-path whereby a full struct is copied from storage to memory, a single member is altered and then the whole struct is once again transmitted from memory to storage, overwriting the previous entry.

Recommendations:
It is possible to account for this case and instead update the “stake” member of the “Snapshot” in-place without reading the whole struct from storage, optimizing the gas cost of the function.