[CHANGED] Certik Audit - Exhibit 16 (round 3)
[INFORMATIONAL] Description:
Instead of reading the “length” variable directly as exposed by the mapped array, the array is copied in full to memory and then the “length” member is read.

Recommendations:
The “length” member can be directly read from storage without copying the whole array to memory. We suggest it is done so to reduce the gas cost of the function significantly.